+++ Inventarios/Controllers/Seguridad/MensajesDelSistemaController.cs
@@ -2,7 +2,7 @@
 using Inventarios.Models.Seguridad;
 using Inventarios.Models.TablasMaestras;
 using Inventarios.services.Seguridad;
-
+using Inventarios.Utils;
 using Microsoft.AspNetCore.Mvc;
 
 namespace Inventarios.Controllers.Seguridad
@@ -12,58 +12,72 @@ namespace Inventarios.Controllers.Seguridad
     public class MensajesDelSistemaController : ControllerBase
     {
         private readonly MensajesDelSistemaService _service;
+        private Validaciones _validaciones;
 
         private List<MensajesDelSistemaDTO>? list;
 
-        public MensajesDelSistemaController(MensajesDelSistemaService service)
+        public MensajesDelSistemaController(MensajesDelSistemaService service , Validaciones validaciones )
         {
             _service = service;
+            _validaciones = validaciones;   
         }
 
         [HttpPost]
         [ActionName("Add")]
-        public Mensaje Add(Mensajesdelsistema obj)
+        public Mensaje Add(Mensajesdelsistema obj, int idusuario, string token)
         {
+            if (_validaciones.ValidarToken(idusuario, token) == false) return null;
+
             Mensaje list = _service.Add(obj);
             return list;
         }
 
-        [HttpDelete("{id}")]
+        [HttpDelete("{id}/{idusuario}/{token}")]
         [ActionName("Delete")]
-        public Mensaje Delete(int id)
+        public Mensaje Delete(int id, int idusuario, string token)
         {
+            if (_validaciones.ValidarToken(idusuario, token) == false) return null;
+
             Mensaje list = _service.Delete(id);
             return list;
         }
 
         [HttpPut]
         [ActionName("Update")]
-        public Mensaje Update(Mensajesdelsistema obj)
+        public Mensaje Update(Mensajesdelsistema obj, int idusuario, string token)
         {
+            if (_validaciones.ValidarToken(idusuario, token) == false) return null;
+
             Mensaje list = _service.Update(obj);
             return list;
         }
 
-        [HttpGet("{id}")]
+        [HttpGet("{id}/{idusuario}/{token}")]
         [ActionName("GetById")]
-        public List<Mensajesdelsistema>? GetById(int id)
+        public List<Mensajesdelsistema>? GetById(int id, int idusuario, string token)
         {
+            if (_validaciones.ValidarToken(idusuario, token) == false) return null;
+
             List<Mensajesdelsistema> list = _service.GetById(id);
             return list;
         }
 
-        [HttpGet("{filtro?}")]
+        [HttpGet("{filtro}/{idusuario}/{token}")]
         [ActionName("GetAll")]
-        public List<MensajesDelSistemaDTO>? GetAll(string filtro = "")
+        public List<MensajesDelSistemaDTO>? GetAll(string filtro, int idusuario, string token)
         {
+            if (_validaciones.ValidarToken(idusuario, token) == false) return null;
+
             list = _service.List(filtro);
             return list;
         }
 
-        [HttpGet]
+        [HttpGet("{idusuario}/{token}")]
         [ActionName("GetAllActive")]
-        public List<MensajesDelSistemaDTO>? GetAllActive()
+        public List<MensajesDelSistemaDTO>? GetAllActive(int idusuario, string token)
         {
+            if (_validaciones.ValidarToken(idusuario, token) == false) return null;
+
             list = _service.ListActive();
             return list;
         }
